python/requirements.txt
numpy==2.2.3
pylsl==1.17.6
pyserial==3.5
